Move TransformQueue to transform/queue.js
